split settings into categories
