+++ packages/keepkey-desktop/package.json
@@ -1,6 +1,6 @@
 {
     "name": "keepkey-desktop",
-    "version": "1.3.4",
+    "version": "1.3.5",
     "author": {
         "name": "KeepKey",
         "email": "support@keepkey.com"
